@@ -4,7 +4,7 @@ import { Routes, RouterModule } from '@angular/router';
 const routes: Routes = [
     {
         path: '',
-        loadChildren: './story/story.module#StoryModule'
+        loadChildren: () => import('./story/story.module').then(m => m.StoryModule)
     }
 ];
 
