@@ -10,8 +10,8 @@ export class ZoomService {
         const width = window.innerWidth;
         const height = window.innerHeight;
         const bbox: DOMRect = <DOMRect>div.getBoundingClientRect();
-        const xDenom: number = width < height ? 5 : 3;
-        const yDenom: number = width < height ? 5 : 5;
+        const xDenom: number = width < height ? 5 : 2.5;
+        const yDenom: number = width < height ? 3 : 5;
         const tx = (width - bbox.width * 2) / xDenom - bbox.x;
         const ty = (height - bbox.height * 2) / yDenom - bbox.y;
         div.classList.add('zoomed');
